# apps/ledgerline-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ledgerline API",
    "description": "Local statement analyzer. Binds 127.0.0.1 and has no authentication: it is a single-user process holding every statement its owner has imported.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string"
          },
          "rowIndexes": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Present on `zero_amount_rows`: the rows that parsed to $0.00."
          }
        }
      },
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "accountId": {
            "type": "string"
          },
          "rawRowId": {
            "type": [
              "null",
              "string"
            ]
          },
          "postedDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "transactionDate": {
            "type": [
              "null",
              "string"
            ]
          },
          "effectiveDate": {
            "type": "string"
          },
          "amountCents": {
            "type": "integer"
          },
          "balanceCents": {
            "type": [
              "null",
              "integer"
            ]
          },
          "currency": {
            "type": "string"
          },
          "descriptionRaw": {
            "type": "string"
          },
          "descriptionNormalized": {
            "type": "string"
          },
          "merchantId": {
            "type": [
              "null",
              "string"
            ]
          },
          "categoryId": {
            "type": [
              "null",
              "string"
            ]
          },
          "categorySource": {
            "type": [
              "null",
              "string"
            ],
            "enum": [
              "seed",
              "rule",
              "llm",
              "user",
              null
            ]
          },
          "isPending": {
            "type": "boolean"
          },
          "isInternalTransfer": {
            "type": "boolean"
          },
          "transferPairId": {
            "type": [
              "null",
              "string"
            ]
          },
          "refundPairId": {
            "type": [
              "null",
              "string"
            ]
          },
          "isExcluded": {
            "type": "boolean"
          },
          "allowsZeroAmount": {
            "type": "boolean"
          },
          "dedupeKey": {
            "type": "string"
          },
          "dedupeKeyVersion": {
            "type": "string"
          },
          "occurrenceIndex": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "accountId",
          "rawRowId",
          "postedDate",
          "transactionDate",
          "effectiveDate",
          "amountCents",
          "balanceCents",
          "currency",
          "descriptionRaw",
          "descriptionNormalized",
          "merchantId",
          "categoryId",
          "categorySource",
          "isPending",
          "isInternalTransfer",
          "transferPairId",
          "refundPairId",
          "isExcluded",
          "allowsZeroAmount",
          "dedupeKey",
          "dedupeKeyVersion",
          "occurrenceIndex",
          "createdAt",
          "updatedAt"
        ]
      },
      "TransactionSearchRow": {
        "type": "object",
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          },
          "hasFinding": {
            "type": "boolean"
          }
        },
        "required": [
          "transaction",
          "hasFinding"
        ]
      },
      "TransactionPage": {
        "type": "object",
        "properties": {
          "rows": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionSearchRow"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "rows",
          "total",
          "limit",
          "offset"
        ]
      },
...
      "TransactionDetail": {
        "type": "object",
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/Transaction"
          },
          "coveringImports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatementImport"
            }
          },
          "rawText": {
            "type": [
              "null",
              "string"
            ]
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionSourceLine"
            }
          }
        },
        "required": [
          "transaction",
          "coveringImports",
          "rawText",
          "sources"
        ]
      },
...
      "TransactionFilter": {
        "type": "object",
        "properties": {
          "accountIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "merchantIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "categoryIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "descriptorsNormalized": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "minAmountCents": {
            "type": "integer"
          },
          "maxAmountCents": {
            "type": "integer"
          },
          "isPending": {
            "type": "boolean"
          },
          "hasFinding": {
            "type": "boolean"
          },
          "includeInternalTransfers": {
            "type": "boolean"
          },
          "includeExcluded": {
            "type": "boolean"
          },
          "q": {
            "type": "string"
          }
        }
      },
      "TransactionBulkChange": {
        "type": "object",
        "properties": {
          "merchantId": {
            "type": [
              "string",
              "null"
            ]
          },
          "categoryId": {
            "type": [
              "string",
              "null"
            ]
          },
          "isInternalTransfer": {
            "type": "boolean"
          },
          "isExcluded": {
            "type": "boolean"
          }
        }
      },
      "TransactionBulkResult": {
        "type": "object",
        "properties": {
          "dryRun": {
            "type": "boolean"
          },
          "matchCount": {
            "type": "integer"
          },
          "updated": {
            "type": "integer"
          },
          "aliasKeysWritten": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "renormalizeJobId": {
            "type": [
              "null",
              "string"
            ]
          },
          "renormalizeJobCoalesced": {
            "type": "boolean"
          }
        },
        "required": [
          "dryRun",
          "matchCount",
          "updated",
          "aliasKeysWritten",
          "renormalizeJobId",
          "renormalizeJobCoalesced"
        ]
      }
    }
  },
  "paths": {
...
    "/api/transactions": {
      "get": {
        "operationId": "listTransactions",
        "summary": "Filter, search and paginate transactions",
        "tags": [
          "transactions"
        ],
        "description": "Internal transfers are excluded unless asked for — a credit-card payment is not spending (spec 6.3). `hasFinding` comes from `finding_evidence` (spec 2.3).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "accountIds",
            "required": false,
            "description": "Comma-separated account ids"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "merchantIds",
            "required": false,
            "description": "Comma-separated merchant ids"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "categoryIds",
            "required": false,
            "description": "Comma-separated category ids"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "from",
            "required": false,
            "description": "Inclusive ISO date"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "to",
            "required": false,
            "description": "Inclusive ISO date"
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "minAmountCents",
            "required": false
          },
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "maxAmountCents",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "isPending",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "hasFinding",
            "required": false
          },
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "includeInternalTransfers",
            "required": false
          },
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "includeExcluded",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": false,
            "description": "Full-text across raw and normalized descriptors"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "date_desc",
                "date_asc",
                "amount_desc",
                "amount_asc"
              ],
              "default": "date_desc"
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionPage"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/{id}": {
      "get": {
        "operationId": "getTransaction",
        "summary": "One transaction, with the imports that cover it",
        "tags": [
          "transactions"
        ],
        "description": "The covering imports and the verbatim line are spec 6.3’s row expander. `rawText` is the statement line as printed, never trimmed or re-encoded (spec 2.5).",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionDetail"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateTransaction",
        "summary": "Assign merchant or category, mark internal transfer, exclude",
        "tags": [
          "transactions"
        ],
        "description": "A merchant assignment writes a `user` merchant_alias and enqueues a coalesced re-normalize job (spec 4.3, 2.7).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionBulkChange"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/bulk": {
      "post": {
        "operationId": "bulkUpdateTransactions",
        "summary": "Apply one change to a filter-matched set",
        "tags": [
          "transactions"
        ],
        "description": "`?dryRun=true` returns the match count only and writes nothing — this is what backs spec 6.3’s \"apply to all 47 matching descriptors\". A real apply of a merchant writes one `user` merchant_alias per matched descriptor (spec 4.3) and enqueues a coalesced re-normalize job (spec 2.7).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "filter"
                ],
                "properties": {
                  "filter": {
                    "$ref": "#/components/schemas/TransactionFilter"
                  },
                  "change": {
                    "$ref": "#/components/schemas/TransactionBulkChange"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "boolean",
              "default": false
            },
            "in": "query",
            "name": "dryRun",
            "required": false,
            "description": "Count the matched set and write nothing"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionBulkResult"
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}